ansible/roles/world_model_service/files/requirements.txt
fastapi
uvicorn
paho-mqtt
httpx
